+++ src/__test__/components/Navbar.test.js
@@ -1,7 +1,7 @@
-import React from 'react';
-import { shallow } from 'enzyme';
-import toJson from 'enzyme-to-json';
-import Navbar from '../../components/Navbar';
+import React from 'react'
+import { shallow } from 'enzyme'
+import toJson from 'enzyme-to-json'
+import Navbar from '../../components/Navbar'
 import {
   NETWORK_GRAPH_VIEW
 } from '../../constants/views'
@@ -12,22 +12,22 @@ const setup = () => {
     setStoreState: jest.fn(),
     mainView: NETWORK_GRAPH_VIEW,
     availableNodes: [{ id: '123' }]
-  };
+  }
 
-  const component = shallow(<Navbar {...props} />);
+  const component = shallow(<Navbar {...props} />)
 
   return {
     component,
     props
-  };
-};
+  }
+}
 
 describe('Navbar', () => {
   it('should match snapshot ', () => {
     const {
       component
-    } = setup();
+    } = setup()
 
-    expect(toJson(component)).toMatchSnapshot();
-  });
-});
+    expect(toJson(component)).toMatchSnapshot()
+  })
+})