+++ src/__test__/components/SearchBox.test.js
@@ -1,21 +1,21 @@
-import React from 'react';
-import { shallow } from 'enzyme';
-import toJson from 'enzyme-to-json';
-import ElementInfo from '../../components/ElementInfo';
+import React from 'react'
+import { shallow } from 'enzyme'
+import toJson from 'enzyme-to-json'
+import ElementInfo from '../../components/ElementInfo'
 
 const setup = () => {
   const props = {
     searchFilter: 'abc',
     setStoreState: jest.fn()
-  };
+  }
 
-  const component = shallow(<ElementInfo {...props} />);
+  const component = shallow(<ElementInfo {...props} />)
 
   return {
     component,
     props
-  };
-};
+  }
+}
 
 describe('ElementInfo', () => {
   afterEach(() => {
@@ -25,8 +25,8 @@ describe('ElementInfo', () => {
   it('should match snapshot ', () => {
     const {
       component
-    } = setup();
+    } = setup()
 
-    expect(toJson(component)).toMatchSnapshot();
-  });
-});
+    expect(toJson(component)).toMatchSnapshot()
+  })
+})