+++ src/__test__/components/Sidebar.test.js
@@ -1,28 +1,28 @@
-import React from 'react';
-import { shallow } from 'enzyme';
-import toJson from 'enzyme-to-json';
-import Sidebar from '../../components/Sidebar';
+import React from 'react'
+import { shallow } from 'enzyme'
+import toJson from 'enzyme-to-json'
+import Sidebar from '../../components/Sidebar'
 
 const setup = () => {
   const props = {
     isSidebarOpen: true,
     setStoreState: jest.fn(),
-  };
+  }
 
-  const component = shallow(<Sidebar {...props} />);
+  const component = shallow(<Sidebar {...props} />)
 
   return {
     component,
     props
-  };
-};
+  }
+}
 
 describe('Sidebar', () => {
   it('should match snapshot ', () => {
     const {
       component
-    } = setup();
+    } = setup()
 
-    expect(toJson(component)).toMatchSnapshot();
-  });
-});
+    expect(toJson(component)).toMatchSnapshot()
+  })
+})