+++ src/__test__/pages/_app.test.js
@@ -1,22 +1,22 @@
-import React from 'react';
-import { shallow } from 'enzyme';
-import toJson from 'enzyme-to-json';
-import MyApp from '../../pages/_app';
-import HeaderComponent from '../../components/HeaderComponent';
+import React from 'react'
+import { shallow } from 'enzyme'
+import toJson from 'enzyme-to-json'
+import MyApp from '../../pages/_app'
+import HeaderComponent from '../../components/HeaderComponent'
 
 const setup = (pageProps) => {
   const props = {
     Component: HeaderComponent,
     pageProps
-  };
+  }
 
-  const component = shallow(<MyApp {...props} />);
+  const component = shallow(<MyApp {...props} />)
 
   return {
     component,
     props
-  };
-};
+  }
+}
 
 describe('_app', () => {
   it('should match snapshot ', () => {
@@ -24,18 +24,18 @@ describe('_app', () => {
       component
     } = setup({
       success: false
-    });
+    })
 
-    expect(toJson(component)).toMatchSnapshot();
-  });
+    expect(toJson(component)).toMatchSnapshot()
+  })
 
   it('should match snapshot if error occurs', () => {
     const {
       component
     } = setup({
       error: { statusCode: 400 }
-    });
+    })
 
-    expect(toJson(component)).toMatchSnapshot();
-  });
-});
+    expect(toJson(component)).toMatchSnapshot()
+  })
+})