+++ src/__test__/pages/index.test.js
@@ -1,19 +1,19 @@
-import React from 'react';
-import { shallow } from 'enzyme';
-import toJson from 'enzyme-to-json';
-import Index from '../../pages/index';
+import React from 'react'
+import { shallow } from 'enzyme'
+import toJson from 'enzyme-to-json'
+import Index from '../../pages/index'
 
 const setup = () => {
-  const component = shallow(<Index />);
+  const component = shallow(<Index />)
 
   return {
     component
-  };
-};
+  }
+}
 
 describe('Index page', () => {
   it('should match snapshot', () => {
-    const { component } = setup();
-    expect(toJson(component)).toMatchSnapshot();
-  });
-});
+    const { component } = setup()
+    expect(toJson(component)).toMatchSnapshot()
+  })
+})