+++ src/components/FooterComponent.js
@@ -1,4 +1,4 @@
-import React from 'react';
+import React from 'react'
 import { connect } from 'redux-zero/react'
 import { useTranslation } from 'react-i18next'
 import PropTypes from 'prop-types'
@@ -47,7 +47,7 @@ const FooterComponent = ({
       </div>
     </footer>
   )
-};
+}
 
 FooterComponent.propTypes = {
   isEdgeFilterOpen: PropTypes.bool.isRequired,