+++ src/components/HeadTags.js
@@ -1,5 +1,5 @@
-import Head from 'next/head';
-import PropTypes from 'prop-types';
+import Head from 'next/head'
+import PropTypes from 'prop-types'
 
 const HeadTags = ({
   title,
@@ -13,16 +13,16 @@ const HeadTags = ({
     <meta name="description" content={`${description}`} />
     <meta name="viewport" content="initial-scale=1.0, width=device-width" />
   </Head>
-);
+)
 
 HeadTags.propTypes = {
   title: PropTypes.string,
   description: PropTypes.string,
-};
+}
 
 HeadTags.defaultProps = {
   title: undefined,
   description: undefined
-};
+}
 
-export default HeadTags;
+export default HeadTags