+++ src/components/Navbar.js
@@ -1,4 +1,4 @@
-import React from 'react';
+import React from 'react'
 import { connect } from 'redux-zero/react'
 import { useTranslation } from 'react-i18next'
 import PropTypes from 'prop-types'
@@ -93,7 +93,7 @@ const Navbar = ({
       </div>
     </nav>
   )
-};
+}
 
 Navbar.propTypes = {
   // isSidebarOpen: PropTypes.bool.isRequired,