+++ CHANGELOG.md
@@ -18,14 +18,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [0.0.2]
 ### Added
-- Added node removal feature
-- Added nodes/edges count
+- Hide edges on drag for performance
+- Network graph loading bar
+- Navigation buttons
+- Node removal feature
+- Nodes/edges count
 
 ### Changed
 
 ### Deprecated
 
 ### Removed
+- Zoom button from settings as already present in navigation buttons
 
 ### Fixed
 
@@ -38,6 +42,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Node search widget
 - Node info widget
 - Node/edges visualisation from JSOn file
+- Internationalisation
 - NextJS boilerplate
 
 ### Changed